+++ src/Explosion.cpp
@@ -48,9 +48,9 @@ Explosion::Explosion(int _x, int _y, int _power, Bomber* _owner, GameContext* co
     }
     
     // Create additional particle effects using ObjectPool pattern
-    ParticleSystem* explosion_sparks = GameObjectFactory::getInstance().create_particle_system(_x, _y, EXPLOSION_SPARKS, get_context());
+    ParticleSystem* explosion_sparks = GameObjectFactory::getInstance().create_particle_system(_x, _y, static_cast<int>(EXPLOSION_SPARKS), get_context());
     
-    ParticleSystem* dust_cloud = GameObjectFactory::getInstance().create_particle_system(_x, _y, DUST_CLOUDS, get_context());
+    ParticleSystem* dust_cloud = GameObjectFactory::getInstance().create_particle_system(_x, _y, static_cast<int>(DUST_CLOUDS), get_context());
 
     length_up = length_down = length_left = length_right = 0;
